refactor: move parser logic to dedicated module
