+++ include/compiler.h
@@ -3,34 +3,19 @@
 
 #include "common.h"
 #include "vm.h"
-#include "lexer.h"
 #include "ast.h"
 
-// Forward declarations
-typedef struct ASTNode ASTNode;
-
-// Compiler functions
 void initCompiler(Compiler* compiler, Chunk* chunk, const char* fileName, const char* source);
 uint8_t allocateRegister(Compiler* compiler);
 void freeRegister(Compiler* compiler, uint8_t reg);
 bool compile(ASTNode* ast, Compiler* compiler, bool isModule);
 
-// Code emission functions
 void emitByte(Compiler* compiler, uint8_t byte);
 void emitBytes(Compiler* compiler, uint8_t byte1, uint8_t byte2);
 void emitConstant(Compiler* compiler, uint8_t reg, Value value);
 
-// Simple parsing function for testing
-ASTNode* parseSource(const char* source);
-void freeAST(ASTNode* node);
+// Compilation helpers
+bool compileExpression(ASTNode* node, Compiler* compiler);
+int compileExpressionToRegister(ASTNode* node, Compiler* compiler);
 
-// Expression parsing functions
-ASTNode* parseExpression();
-ASTNode* parseBinaryExpression(int minPrec);
-ASTNode* parsePrimaryExpression();
-
-// Helper functions
-int getOperatorPrecedence(TokenType type);
-const char* getOperatorString(TokenType type);
-
-#endif
+#endif // COMPILER_H
